Move validation utils to their own submodule
